+++ src/view/components/InspectButton.svelte
@@ -1,6 +1,25 @@
 <script>
-   export let item;
    import InlineButton from "crew-components/InlineButton";
+   import { onHook, calculateValue, logger } from "crew-components/helpers";
+   import { tick } from "svelte";
+
+   export let item;
+   export let size;
+   const isString = (i) => i instanceof String || typeof i === "string";
+
+   function updateItem() {
+      if (item && isString(item)) {
+         try {
+            item = calculateValue(item, "actor");
+            if (item) {
+               item = { source: item };
+            }
+         } catch (_) {
+            item = null;
+         }
+      }
+     return item;
+   }
 
    async function inspectItem(e, item) {
       e.stopPropagation();
@@ -10,6 +29,16 @@
       }
       globalThis.game.modules.get("data-inspector")?.api?.inspect(doc);
    }
+   updateItem();
+   onHook(["updateActor", "controlToken", "targetToken"], _ => tick().then(updateItem));
 </script>
 
-<InlineButton title="Inspect" color="#71717a" icon="fa6-solid:magnifying-glass" on:click={(e) => inspectItem(e, item)} />
+{#if item && game.modules.get("data-inspector")?.api}
+   <InlineButton
+      title="Inspect"
+      color="#71717a"
+      icon="fa6-solid:magnifying-glass"
+      on:click={(e) => inspectItem(e, updateItem())}
+      {size}
+   />
+{/if}